Try to adapt Santana's PHA learning to NES jump learning.
Run `python modelearner.py`.  Note that the initial guess for the system
defined by `jump_data()` is not too bad, but the first maximization step
breaks terribly with a singular matrix.  I am not sure whether the issue
is in the A and B matrices or the weights (gamma).

This could be caused either by bad input data (I tried to mimic the
definition of the lawnmower HA in mm_systems.py, but using the emulator
as the source of positions rather than a model) or by the HA being too
nonlinear or something for the PHA learning code to learn.

I think bad input data is more likely; I tried a few different values
for `sigmaLik` which was initially 0.005 or something, leading to 0
log-likelihood for any sequence of Y-values.  Increasing that number
seemed to account better for imprecision in the inputs.

Apologies for weird code formatting, autopep8 sometimes acts badly.

Note: jump_data currently only uses a few trials and concatenates them;
it also assumes perfect knowledge of what mode the character is in.