PyMC3 segmented regression with known segment boundaries.
Seems to work for finding jump accelerations for Mario and Metroid,
though I have not fully closed the loop by identifying hard VY resets
and updating HA parameters and checking errors in the abstraction and
so on.  Note that I know basically nothing about what I'm doing
statistics-wise, so I'm eagerly awaiting feedback on this.

I also would like some info on making a nice visualization for this.
I had a thing from sample code for visualizing a single bundle of
likely lines from a model, but it was quite specific to that sample
code.  I'd like to use the GLM module's plotting but it seems hard-
coded to use pre-defined variable names and I also can't see how
to get the piecewise stuff working except by repeated plotting (which
would be fine, it's just work).