Finally fixed Mario jumping.  The automaton was wrong.
Added distinct "up-control" and "up-fixed" state, where up-control
states describe the continued influence of the jump button.  In both
up states, gravity is lesser for Mario than in the down state.  For
characters like Samus without distinct up-control and up-fixed states,
we act as if we always pass through up-fixed on the way to down when
releasing the button.  This learns quite uncertain _flows_ on
up-fixed, though the _resets_ are fairly good and precise.