Pretty much got segmentation working, though the segments are bad.
This may be because of the wrong criterion, because of comparing
criteria across multiple models on not-identical data (in the
minimization part), or because of small sample sizes in some
intervals, or because of some other issue.

I guess it's also possible we aren't getting the right transition
instants.

For now I just use max-likelihood as the criterion because it seemed
least likely to explode with extremely large or small values.