Got Adam's stuff reproduced.  Draft NPMI cause learning.
Interpreting the results is a bit tricky, but I think this is moving
in the right direction.  Now we learn that velocity going from
negative to 0 is strongly associated with entering the high-gravity
falling state from the jumping state (I think).  Also pressing A and
having velocity go negative are associated with jumping in general,
although specific jump transitions are more associated with what tiles
you happen to be touching at the time (seems undesirable but understandable.)