Commit not-quite-working guard calculation.
The basic idea seems OK but the data are funky.  A button presses are
not being seen at the expected NPMI with "entering the jump state", so
maybe there's some offset or some other problem going on.  Also, the
grouped collision events are not coming up with as high an NPMI as we
might hope for entering jumping, the individual events are looking
more likely.