Textual HA output.  Note units are pixels per second.
Or pixels per second squared!  The textual format should be pretty
good for processing in something like awk or sed or perl or python or
something.

Maybe worth factoring out some of the if __name__ == "__main__" stuff
in jump_from_tracks.py to a new function you can call and drive from a
directory enumeration or whatever.  Or write a shell script!

It would be an easy-ish tweak to figure out which button jumps, but
for now we just pass it in as an argument.